+++ Neobot iOS Client/Neobot iOS Client/PXGTools.h
@@ -12,8 +12,8 @@
 #import <UIKit/UIKit.h>
 
 
-NSDictionary* pxgEncodePointData(int x, int y, double theta);
-void pxgDecodePointData(NSDictionary* dictionnary, int* x, int* y, double* theta);
+NSDictionary* pxgEncodePointData(double x, double y, double theta);
+void pxgDecodePointData(NSDictionary* dictionnary, double* x, double* y, double* theta);
 
 NSDictionary* pxgEncodeTrajectoryData(NSString* name, NSArray* pointsData);
 void pxgDecodeTrajectoryData(NSDictionary* dictionnary, NSString** name, NSArray** pointsData);
